apps/api/pyproject.toml
[project]
name = "bimstitch-api"
version = "0.0.1"
description = "BIMstitch FastAPI backend with auth"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "fastapi-users[sqlalchemy]>=14.0.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "pydantic[email]>=2.9.0",
    "python-jose[cryptography]>=3.3.0",
    "aiosmtplib>=3.0.2",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.2",
    "redis>=5.0.0",
    "fastapi-limiter>=0.1.6,<0.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "asgi-lifespan>=2.1.0",
    "ruff>=0.7.4",
    "mypy>=1.13.0",
    "types-python-jose>=3.3.4.20240106",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bimstitch_api"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "F", "W", "I", "B", "UP", "N", "SIM", "TCH", "RUF",
]
ignore = ["B008"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
packages = ["bimstitch_api"]
mypy_path = "src"

[[tool.mypy.overrides]]
module = ["fastapi_users.*", "aiosmtplib.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
pythonpath = ["src"]
filterwarnings = [
    "ignore::DeprecationWarning",
]
